Modification: Statistics
Now only the triplets sets is necesary to provide the statistics of the
dataset. If the entity set and relationship set are in Wikidata format,
additional information will be autanatically provided (i.e.,
categories). If the node_data and relation_data are provided, the
plotting can be used to show the rankings and names.